resnet-classifier/requirements.txt
flask==3.0.0
transformers==4.35.0
pillow==10.1.0
torch==2.1.0
numpy<2
